--- src/Tests/CSVTests/__init__.py
@@ -1,3 +1,0 @@
-import unittest
-from CsvReader import csvReader
-from CsvReader import ClassFactory
