+++ src/Tests/CSVTests/test_CSV.py
@@ -1,6 +1,6 @@
 import unittest
-from CsvReader import csvReader
-from CsvReader import ClassFactory
+from src.CsvReader.csvReader import CsvReader as ReadCSV
+from src.CsvReader.ClassFactory import class_factory
 
 
 class MyTestCase(unittest.TestCase):
@@ -11,7 +11,7 @@ class MyTestCase(unittest.TestCase):
     def test_return_data_as_objects(self):
         values = self.csv_reader.return_data_as_objects('a_value')
         self.assertIsInstance(values, list)
-        test_class = ClassFactory.class_factory('a_value', self.csv_reader.data[0])
+        test_class = class_factory('a_value', self.csv_reader.data[0])
         for a_value in values:
             self.assertEqual(a_value[0], test_class[0])
