--- src/Tests/test_Calculator/__init__.py
@@ -1,4 +1,0 @@
-import unittest
-from Calculator import Calculator
-from CsvReader import CsvReader
-from decimal import Decimal
